Multi Stage race dev started
If the condition of the initial if statement is not true, it means that it is a multi-stage race. In this case, the code inside the else block is executed.

Inside the else block, there is another if statement that compares the racers.number with the data[i+1] value, but with the last character removed using [:-1]. This is done to match all the numbers where the last digit is always the special number.

If there is a match, the special number is appended to the racers.stage list. Additionally, based on the file.filename, the corresponding start or finish time is appended to the racers.start or racers.finish lists, respectively. It is assumed that these attributes are initially empty lists.

Finally, the racers.calculate_time() method is called to calculate the total time for the racer.

In summary, this code snippet handles race data for both single stage and multi-stage races. It looks for matching numbers in the data and performs different actions based on the race stage and file type. It also calculates the total time for each racer.